requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
google-generativeai>=0.3.0
langchain-google-genai>=1.0.0
chromadb>=0.4.15
crewai>=0.1.0
langchain>=0.0.334
mlflow>=2.8.0
python-multipart>=0.0.6
pillow>=10.0.1
numpy>=1.24.3
pandas>=2.0.3
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
aiofiles>=23.2.1
pypdf2>=3.0.1
python-docx>=0.8.11
chardet>=5.2.0